move cycle_scalar to new files
